+++ RELEASE_NOTES.md
@@ -1,3 +1,9 @@
 ### New Features
 
-- Implement Subscribers API
+Adding support for the Widget/Content endpoints. See the [Widgets Tutorial](https://developer.bigcommerce.com/api-docs/store-management/widgets/tutorials/tutorial)
+for usage.
+
+Implemented:
+ - [Widgets API](https://developer.bigcommerce.com/api-reference/store-management/widgets/widget/createwidget)
+ - [Widget Templates](https://developer.bigcommerce.com/api-reference/store-management/widgets/widget-template/createwidgettemplate)
+ - Get Theme Regions